attack range attribute tweaks

+++ src/main/java/com/github/theredbrain/minecrawl/entity/DuckLivingEntityMixin.java
@@ -1,5 +1,5 @@
 package com.github.theredbrain.minecrawl.entity;
 
 public interface DuckLivingEntityMixin {
-	int minecrawl$getAttackRange();
+	double minecrawl$getAttackRange();
 }
